requirements.txt
matplotlib
Pillow
joblib
faiss-cpu
scikit-learn
config
torch
torchvision
numpy